conditional websocket setup if websocket is not plugin - Add install.bat after folder was created - Remove duplicate download

+++ Install.ps1
@@ -27,12 +27,6 @@ else {
     Write-Host "process obs64.exe not found."
 }
 
-# Загружаем последний релиз AutoRecord.zip с GitHub
-Write-Host "Downloading latest release of AutoRecord..."
-Invoke-WebRequest -Uri $DOWNLOAD_URL -OutFile $ZIP_FILE -UseBasicParsing
-
-(Invoke-WebRequest https://raw.githubusercontent.com/TheLeonelus/AutoRecord/main/install.bat).Content | Out-File -FilePath $AUTO_RECORD_PATH\install.bat -Encoding UTF8
-
 # Clean old code
 if (Test-Path -Path $AUTO_RECORD_PATH) {
     if (Test-Path -Path $AUTO_RECORD_PATH\Lib\) {
@@ -48,7 +42,8 @@ if (Test-Path -Path $AUTO_RECORD_PATH) {
 Write-Host "Downloading latest release of AutoRecord..."
 try {
     Invoke-WebRequest -Uri $DOWNLOAD_URL -OutFile $ZIP_FILE -UseBasicParsing
-} catch {
+}
+catch {
     Write-Host "Failed to download AutoRecord.zip. Exiting."
     Read-Host -Prompt "Press Enter to exit"
     exit 1
@@ -58,7 +53,11 @@ try {
 Expand-Archive -Path $ZIP_FILE -DestinationPath $AUTO_RECORD_PATH -Force
 Remove-Item -Path $ZIP_FILE
 
+# Download install.bat to be able to update in the future
+(Invoke-WebRequest https://raw.githubusercontent.com/TheLeonelus/AutoRecord/main/install.bat).Content | Out-File -FilePath $AUTO_RECORD_PATH\install.bat -Encoding UTF8
+
 # Import obs-websocket config
+Write-Host "Inserting default websocket config"
 $jsonConfig = '
 {
     "alerts_enabled":  false,
@@ -68,27 +67,44 @@ $jsonConfig = '
     "server_password":  "",
     "server_port":  4455
 }' | ConvertFrom-Json
-$jsonConfig | ConvertTo-Json | Set-Content -Path "$env:appdata\obs-studio\plugin_config\obs-websocket\config.json" -Encoding UTF8
-Write-Host "Inserting default websocket config"
+if (Test-Path -Path "$env:appdata\obs-studio\plugin_config\obs-websocket\") {
+    $jsonConfig | ConvertTo-Json | Set-Content -Path "$env:appdata\obs-studio\plugin_config\obs-websocket\config.json" -Encoding UTF8
+}
+else {
+    $global = "$env:appdata\obs-studio\global.ini"
+    $regex = '(?ms)(\[OBSWebSocket\]\n)(.*?)(\n\[)'
+    (Get-Content $global -Raw -Encoding UTF8) -replace $regex, 
+    '$1FirstLoad=false
+ServerEnabled=true
+ServerPort=4455
+AlertsEnabled=false
+AuthRequired=false
+ServerPassword=
+$3' | Set-Content $global -Encoding UTF8
+}
+
 
 # Create shortcut on desktop
 $ws = New-Object -ComObject WScript.Shell
+Write-Host "Adding AutoRecord to Desktop"
 $shortcut = $ws.CreateShortcut("$DESKTOP_PATH\AutoRecord.lnk")
 $shortcut.TargetPath = $AUTO_RECORD_EXE
 $shortcut.Save()
-Write-Host "Adding AutoRecord to Desktop"
+
 
 # Create AutoRecord shortcut in startup
+Write-Host "Adding AutoRecord to Startup"
 $shortcut = $ws.CreateShortcut("$STARTUP_PATH\AutoRecord.lnk")
 $shortcut.TargetPath = $AUTO_RECORD_EXE
 $shortcut.Save()
-Write-Host "Adding AutoRecord to Startup"
+
 
 # Create OBS shortcut in startup
+Write-Host "Adding OBS to Startup"
 $shortcut = $ws.CreateShortcut("$STARTUP_PATH\obs64.lnk")
 $shortcut.TargetPath = ("C:\Program Files\obs-studio\bin\64bit\obs64.exe")
 $shortcut.WorkingDirectory = ("C:\Program Files\obs-studio\bin\64bit")
 $shortcut.Save()
-Write-Host "Adding OBS to Startup"
+
 
 Read-Host -Prompt "Press Enter to exit"
